+++ docsrc/demos/CustomTextField.demo.js
@@ -1,0 +1,18 @@
+import React, { useState } from "react";
+import MultiPicker from "../../src/index";
+import { getSuggestedFruitSync } from "./common";
+
+export default function CustomTextFieldDemo() {
+    const [items, setItems] = useState([]);
+    return (
+        <MultiPicker
+            value={ items }
+            onChange={ setItems }
+            itemToString={ fruit => fruit.name }
+            getSuggestedItems={ getSuggestedFruitSync }
+            label="Your favourite fruit"
+            fullWidth
+            variant="outlined"
+        />
+    );
+}